# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Articles API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/get-token": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Get a token for API access",
                "operationId": "1ffa1b88745ab6411e6099ca84b90a10",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "device_name"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "admin"
                                    },
                                    "device_name": {
                                        "type": "string",
                                        "example": "react"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token successfully generated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized access",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "incorrect_access_data"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/article": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": " Get list of articles",
                "operationId": "8ade913446e8385bdeb38f1449c7ee5d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Filter by article ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "Filter by user ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Filter by article title",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of articles"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    },
                    {
                        "customAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Articles"
                ],
                "summary": "Create a new article",
                "operationId": "25e58c50f92df67696a665328c72ef42",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Required if article does not exist",
                                        "type": "string",
                                        "maxLength": 64
                                    },
                                    "content": {
                                        "type": "string",
                                        "maxLength": 500,
                                        "nullable": false
                                    },
                                    "article_cover_img": {
                                        "description": "Image file in jpeg, jpg, or png format",
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Article created"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    },
                    {
                        "customAuth": []
                    }
                ]
            }
        },
        "/api/article/{article}": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get a article by ID",
                "operationId": "562e7df204bee3cc349648a6b3b0b5d5",
                "parameters": [
                    {
                        "name": "article",
                        "in": "path",
                        "description": "ID of the article",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article details"
                    },
                    "404": {
                        "description": "Article not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    },
                    {
                        "customAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Articles"
                ],
                "summary": "Update an article",
                "operationId": "c88f4c3179fefd6936a3e872b28db905",
                "parameters": [
                    {
                        "name": "article",
                        "in": "path",
                        "description": "ID of the article",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "content"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Required if article does not exist",
                                        "type": "string",
                                        "maxLength": 64
                                    },
                                    "content": {
                                        "description": "Required if article does not exist",
                                        "type": "string",
                                        "maxLength": 500
                                    },
                                    "article_cover_img": {
                                        "description": "Image file in jpeg, jpg, or png format",
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    },
                                    "_method": {
                                        "description": "Simulate a PUT request",
                                        "type": "string",
                                        "example": "PUT"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Article updated"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Article not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    },
                    {
                        "customAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Articles"
                ],
                "summary": "Delete a article",
                "operationId": "b31b56ebb443d613b08bd22f7e186006",
                "parameters": [
                    {
                        "name": "article",
                        "in": "path",
                        "description": "ID of the article",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article deleted"
                    },
                    "404": {
                        "description": "Article not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    },
                    {
                        "customAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "customAuth": {
                "type": "apiKey",
                "description": "Enter token in format  Bearer xx|xxxxxxxx",
                "name": "Authorization",
                "in": "header",
                "scheme": "Bearer"
            },
            "sanctum": {
                "type": "http",
                "description": "Enter token in format without word Bearer",
                "name": "Authorization",
                "in": "header",
                "scheme": "Bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Authentication",
            "description": "Authentication"
        },
        {
            "name": "Articles",
            "description": "Articles"
        }
    ]
}